Referencing the below article to build modular game structure
[this](https://replit.com/@mkihanya33/ConnectFourWithDOMSkeleton?replId=0320e2a3-690d-4968-9c2b-44f4db614552#script.js)

Next:
- determine method to represent board in array
so that user input can access the specific position
in the array that can change. Not every position in the
board array should be changeable. I only want certain
positions to be able to change and have those positions
be accessible and editable through user input (the
user input will happen at some point).